// src/types/utools.d.ts
export interface TplFeatureArgs {
  /**
   * 进入插件时调用（可选）
   */
  enter?: TplFeatureArgsEnter
  /**
   * 子输入框内容变化时被调用 可选 (未设置则无搜索)
   */
  search?: TplFeatureArgsSearch
  /**
   * 用户选择列表中某个条目时被调用
   */
  select?: TplFeatureArgsSelect
  /**
   * 子输入框为空时的占位符，默认为字符串"搜索"
   */
  placeholder: string

  /**
   * 文档模式的索引集合  indexes: require('./indexes.json')
   */
  indexes?: Array<{
    /**
     * 标题
     */
    t: string

    /**
     * 描述
     */
    d: string

    /**
     * 页面, 只能是相对路径
     */
    p: string
  }>
}

export type TplFeatureMode = 'list' | 'doc' | 'none'

/**
 * @description 模板插件 Feature
 */
export interface TplFeature {
  mode: TplFeatureMode
  args: TplFeatureArgs
}

/**
 * @description 模板插件, feature-code: Feature
 */
export interface TemplatePlugin {
  [index: string]: TplFeature
}